+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/sql/RunAsRole.java
@@ -1,0 +1,16 @@
+// Code generated from OpenAPI specs by Databricks SDK Generator. DO NOT EDIT.
+
+package com.databricks.sdk.service.sql;
+
+import com.databricks.sdk.support.Generated;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+/** Run as role */
+@Generated
+public enum RunAsRole {
+  @JsonProperty("owner")
+  OWNER,
+
+  @JsonProperty("viewer")
+  VIEWER,
+}